chore: add gemini + vertex to new agent loop  (#2230)
