+++ tests/configs/llm_model_configs/gemini-2.5-pro-vertex.json
@@ -1,5 +1,5 @@
 {
-  "model": "gemini-2.5-pro-exp-03-25",
+  "model": "gemini-2.5-pro-preview-05-06",
   "model_endpoint_type": "google_vertex",
   "model_endpoint": "https://us-central1-aiplatform.googleapis.com/v1/projects/memgpt-428419/locations/us-central1",
   "context_window": 1048576,
